removed the isValidSelection method from DeleteAction and MultipleSelectionAction to AFeatureModelAction
